chore: integrated jwt token auth

+++ backend/src/services/posts.js
@@ -1,7 +1,8 @@
 import { Post } from '../db/models/post.js'
+import { User } from '../db/models/user.js';
 
-export async function createPost({ title, content, author, tags }) {
-  const post = new Post({ title, content, author, tags })
+export async function createPost(userId, { title, content, tags }) {
+  const post = new Post({ title, author: userId, content, tags})
   return await post.save()
 }
 
@@ -9,8 +10,10 @@ export async function listAllPosts(options) {
   return await listPost({}, options)
 }
 
-export async function listPostsByAuthor(author, options) {
-  return await listPost({ author }, options)
+export async function listPostsByAuthor(authorUsername, options) {
+  const user = await User.findOne({ username: authorUsername });
+  if (!user) return [];
+  return await listPost({ author: user._id }, options);
 }
 
 export async function listPostsByTag(tag, options) {
@@ -21,16 +24,16 @@ export async function getPostById(postId) {
   return await Post.findById(postId)
 }
 
-export async function updatePost(postId, { title, content, author, tags }) {
+export async function updatePost(userId, postId, { title, content, tags }) {
   return await Post.findByIdAndUpdate(
-    { _id: postId },
-    { $set: { title, content, author, tags } },
+    { _id: postId, author: userId },
+    { $set: { title, content, tags } },
     { new: true },
   )
 }
 
-export async function deletePost(postId) {
-  return await Post.deleteOne({ _id: postId })
+export async function deletePost(userId, postId) {
+  return await Post.deleteOne({ _id: postId, author: userId })
 }
 
 async function listPost(
